fix icon color change(refactoring), add theme_settings_widget.dart
