+++ lib/modules/style_manager/bloc/change_theme_bloc/change_theme_bloc.dart
@@ -1,4 +1,3 @@
-import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:chat_app/modules/storage_manager/preferences/user_preferences.dart';
 import 'package:chat_app/modules/style_manager/themes/custom_themes.dart';
@@ -7,7 +6,6 @@ import 'change_theme_event.dart';
 import 'change_theme_state.dart';
 
 class ChangeThemeBloc extends Bloc<ChangeThemeEvent, ChangeThemeState> {
-  // final bloc = ChangeThemeBloc();
   ChangeThemeBloc()
       : super(ChangeThemeState(
           theme: CustomTheme.darkThemeDeepPurple,
@@ -18,14 +16,10 @@ class ChangeThemeBloc extends Bloc<ChangeThemeEvent, ChangeThemeState> {
 
   void _setThemeEvent(
       SetThemeEvent event, Emitter<ChangeThemeState> emit) async {
-    // initState();
     await UserPreferences().SetThemeEvent('lightThemeDeepPurple');
     await UserPreferences().getTheme();
-    print(event.index);
-    print('state.index ${state.iconColor[event.index]}');
 
     emit(ChangeThemeState(
         theme: CustomTheme().themes[event.index]!, index: event.index));
-    // dispose();
   }
 }
